requirements.txt
APScheduler==3.6.3
asgiref==3.5.0
attrs==21.4.0
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.10
cssselect2==0.6.0
Django==4.0.3
django-cors-headers==3.12.0
django-crispy-forms==1.14.0
django-filter==21.1
djangorestframework==3.13.1
easy-thumbnails==2.8.1
fontawesomefree==6.1.1
idna==3.3
iniconfig==1.1.1
lxml==4.8.0
packaging==21.3
Pillow==9.1.0
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pyTelegramBotAPI==4.4.0
pytest==7.1.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
reportlab==3.6.9
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
tinycss2==1.1.1
tomli==2.0.1
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
webencodings==0.5.1
